Since idQueue only have one thread to add ids from id_queue into database, now we have 60 threads add ids to id_queue,  one thread is too slow, the linked queue will became so huge that it consume too much memory.
Solution: each thread add ids to database directly. Though it might cause duplicated documents, but it only
happened when threads want to insert the same user at the same time, since all the thread with check the user exist before
doing the insert operation, it rarely happened. Therefore, this solution is acceptable.

Note: There are better way to avoid duplicated documents, that is to make the id field in user_profiles collection to be unique,
but mongodb didn't allow us to modified field attributed if the table already have duplicated data.